% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_effective_n}
\alias{calc_effective_n}
\title{Calculate the effective sample size of a case-control GWAS}
\usage{
calc_effective_n(N_case, N_control)
}
\arguments{
\item{N_case}{number of cases}

\item{N_control}{Number of controls}
}
\value{
a double
}
\description{
Calculate the effective sample size of a case-control GWAS
}
\examples{
calc_effective_n(500, 1000)
}
