% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate some metrics from reconstructed and observed streamflow}
\usage{
calculate_metrics(Qa.hat, Qa, z)
}
\arguments{
\item{Qa.hat}{reconstructed}

\item{Qa}{observed}

\item{z}{indices of left-out values in cross validation}
}
\description{
Calculate some metrics from reconstructed and observed streamflow
}
