% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{combine_ld_data}
\alias{combine_ld_data}
\title{Merge two sets of LD data, and save to parquet}
\usage{
combine_ld_data(list1, list2, outdir)
}
\arguments{
\item{list1}{output of \code{\link[=ldsc_to_parquet]{ldsc_to_parquet()}} or \code{\link[=parse_parquet_dir]{parse_parquet_dir()}}}

\item{list2}{output of \code{\link[=ldsc_to_parquet]{ldsc_to_parquet()}} or \code{\link[=parse_parquet_dir]{parse_parquet_dir()}}}

\item{outdir}{directory to store merged LD data}
}
\description{
Merge two sets of LD data, and save to parquet
}
\examples{
\dontrun{
combe_ld_data(l1, l2, "path/to/storage")
}
}
