% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr}
\alias{corr}
\title{Pearson's correlation}
\usage{
corr(x, y)
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}
}
\value{
Pearson's correlation
}
\description{
Calculate the Pearson's correlation using the numerically stable formulation (see References). Internal function.
}
\section{Reference John D. Cook's article at https}{
//www.johndcook.com/blog/2008/11/05/how-to-calculate-pearson-correlation-accurately/
}

