% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{cvLDS}
\alias{cvLDS}
\title{Cross validate LDS model}
\usage{
cvLDS(Qa, u, v, method = "EM", init = NULL, num.restart = 20,
  lambda = 1, k, n.reps = 100, niter = 1000, tol = 1e-05,
  Z = NULL, reupdate = FALSE, parallel = F)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{method}{Either EM or GA (note: GA is experimental)}

\item{init}{A list of initial conditions, each element is a vector of length 4, the initial values for A, B, C and D. The initial values for Q and R are always 1, and mu_1 is 0 and V_1 is 1.}

\item{niter}{Maximum number of iterations, default 1000}

\item{tol}{Tolerance for likelihood convergence, default 1e-5. Note that the log-likelihood is normalized by dividing by the number of observations.}

\item{parallel}{If TRUE, the computation is done in parallel using all available cores
(using the doParallel backend). If FALSE, the computation is done serially.}
}
\description{
Cross validate LDS model
}
\details{
Allows different experimental setups:
\itemize{
\item init: if given, all cross validation runs start with the same init, otherwise each cross validation run is learned using randomized restarts
\item Z: a list of n.reps elements, each is a vector of length k. If given, cross validation will be run on these points (useful when comparing LDS with another reconstruction method); otherwise, randomized cross validation points will be created
}
}
