% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{cvLDS_ensemble}
\alias{cvLDS_ensemble}
\title{Cross validatdion for ensemble reconstruction. This is a wrapper for \link{cvLDS}.}
\usage{
cvLDS_ensemble(Qa, u.list, v.list, start.year, method = "EM",
  trans = "log", k, CV.reps = 100, Z = NULL, init = NULL,
  num.restarts = 100, lambda = 1, ub, lb, num.islands = 4,
  pop.per.island = 100, niter = 1000, tol = 1e-05, parallel = TRUE,
  all.cores = FALSE)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{u.list}{List of u matrices}

\item{v.list}{List of v matrices#'}

\item{start.year}{Starting year of the paleo period. \code{start.year + ncol(u) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{method}{Either EM or GA (note: GA is experimental)}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{k}{Numer of data points to be left out in each cross validation-run.}

\item{CV.reps}{Number of cross-validation runs.}

\item{Z}{A list of CV.reps elements, each is a vector of length k. See Details.}

\item{init}{A list of initial conditions, each element is a vector of length 4, the initial values for A, B, C and D. The initial values for Q and R are always 1, and mu_1 is 0 and V_1 is 1.}

\item{num.restarts}{if init is not given then num.restarts must be provided. In this case the function
will randomize the initial value by sampling uniformly within the range for each parameters
(A in [0, 1], B in [-1, 1], C in [0, 1] and D in [-1, 1]).}

\item{ub}{Upper bounds, a vector whose length is the number of parameters}

\item{lb}{Lower bounds}

\item{niter}{Maximum number of iterations, default 1000}

\item{tol}{Tolerance for likelihood convergence, default 1e-5. Note that the log-likelihood is normalized by dividing by the number of observations.}

\item{parallel}{If TRUE, the computation is done in parallel using all available cores
(using the doParallel backend). If FALSE, the computation is done serially.}
}
\value{
same as cvLDS
}
\description{
Cross validatdion for ensemble reconstruction. This is a wrapper for \link{cvLDS}.
}
