% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cvPCR}
\alias{cvPCR}
\title{Cross validation of PCR reconstruction.}
\usage{
cvPCR(Qa, pc, start.year, trans = "log", k, CV.reps = 100, Z = NULL)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc}{Selected principal components. No input selection is done during cross-validation.}

\item{start.year}{Starting year of the paleo period. \code{start.year + ncol(u) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{k}{Numer of data points to be left out in each cross validation-run.}

\item{CV.reps}{Number of cross-validation runs.}

\item{Z}{A list of CV.reps elements, each is a vector of length k. See Details.}
}
\value{
A list of cross validation results
\itemize{
\item metrics: performance metrics
\item metrics.dist: distribution of performance metrics across all cross-validation runs.
\item Ycv: the predicted streamflow in each cross validation run
\item Z: the cross-validation points
}
}
\description{
Cross validation of PCR reconstruction.
}
