% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cvPCR_ensemble}
\alias{cvPCR_ensemble}
\title{Cross validation of PCR_ensemble reconstruction.}
\usage{
cvPCR_ensemble(
  Qa,
  pc.list,
  start.year,
  trans = "log",
  k,
  CV.reps = 100,
  Z = NULL
)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc.list}{A list, each element is a set of principal component as in \code{PCR_reconstruction}'s \code{pc}}

\item{start.year}{Starting year of the paleo period. \code{start.year + ncol(u) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{k}{Numer of data points to be left out in each cross validation-run.}

\item{CV.reps}{Number of cross-validation runs.}

\item{Z}{A list of CV.reps elements, each is a vector of length k. See Details.}
}
\value{
A list of cross validation results
\itemize{
\item metrics: performance metrics
\item metrics.dist: distribution of performance metrics across all cross-validation runs.
\item Ycv: the predicted streamflow in each cross validation run
\item Z: the cros
}
}
\description{
Cross validation of PCR_ensemble reconstruction.
}
