% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exp_ci}
\alias{exp_ci}
\title{Exponential confidence interval}
\usage{
exp_ci(y, sigma2)
}
\arguments{
\item{y}{A vector of model estimates}

\item{sigma2}{The variance of y as learned from a model}
}
\value{
A data.table with the updated confidence intervals
}
\description{
Get the confidence interval of Q = exp(Y) when Y is normal, i.e, Q is log-normal.
}
