% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_tidyGWAS.R
\name{from_tidyGWAS}
\alias{from_tidyGWAS}
\title{Parse GWAS format of tidyGWAS()}
\usage{
from_tidyGWAS(tbl, n = c("N", "EffectiveN"))
}
\arguments{
\item{tbl}{a \code{\link[dplyr:reexports]{dplyr::tibble()}}}

\item{n}{Column name of sample size, default is "N".}
}
\value{
a munged \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Parse GWAS format of tidyGWAS()
}
\examples{
\dontrun{
munged <- from_tidyGWAS("path/tidyGWAS/cleaned/tidyGWAS_hivestyle")
}
}
