% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_h2.R
\name{get_annot_names}
\alias{get_annot_names}
\title{Print names of annotations in a ldscore fileset}
\usage{
get_annot_names(ldscore_dir)
}
\arguments{
\item{ldscore_dir}{A filepath to a ldscore directory}
}
\value{
a character vector of column names
}
\description{
Print names of annotations in a ldscore fileset
}
\examples{
\dontrun{
get_annot_names("ldscore/dir/celltypes")
}
}
