% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KGE}
\alias{KGE}
\title{Kling-Gupta Efficiency}
\usage{
KGE(yhat, y)
}
\arguments{
\item{yhat}{Model outputs}

\item{y}{Observations}
}
\value{
KGE
}
\description{
Kling-Gupta Efficiency
}
\examples{
KGE(rnorm(100), rnorm(100))
}
