% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{LDS_ensemble}
\alias{LDS_ensemble}
\title{Ensemble reconstruction. This is a wrapper for \link{LDS_reconstruction}.}
\usage{
LDS_ensemble(Qa, u.list, v.list, method = "EM", trans = "log",
  init = NULL, num.restarts = 100, return.init = TRUE, lambda = 1,
  ub = NULL, lb = NULL, num.islands = 4, pop.per.island = 250,
  niter = 1000, tol = 1e-05, return.raw = FALSE, parallel = TRUE)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{u.list}{List of u matrices}

\item{v.list}{List of v matrices#'}

\item{method}{Either EM or GA (note: GA is experimental)}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{init}{A list of initial conditions, each element is a vector of length 4, the initial values for A, B, C and D. The initial values for Q and R are always 1, and mu_1 is 0 and V_1 is 1.}

\item{num.restarts}{if init is not given then num.restarts must be provided. In this case the function
will randomize the initial value by sampling uniformly within the range for each parameters
(A in [0, 1], B in [-1, 1], C in [0, 1] and D in [-1, 1]).}

\item{return.init}{Indicate whether the initial condition that results in the highest
log-likelihood is returned. Default is TRUE.}

\item{ub}{Upper bounds, a vector whose length is the number of parameters}

\item{lb}{Lower bounds}

\item{niter}{Maximum number of iterations, default 1000}

\item{tol}{Tolerance for likelihood convergence, default 1e-5. Note that the log-likelihood is normalized by dividing by the number of observations.}

\item{return.raw}{If TRUE, state and streamflow estimates without measurement updates will be returned.}

\item{parallel}{If TRUE, the computation is done in parallel using all available cores
(using the doParallel backend). If FALSE, the computation is done serially.}
}
\value{
A list of models in the ensemble
}
\description{
Ensemble reconstruction. This is a wrapper for \link{LDS_reconstruction}.
}
