% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_GA.R
\name{LDS_GA}
\alias{LDS_GA}
\title{Learn a linear dynamical system using Genetic Algorithm}
\usage{
LDS_GA(y, u, v, lambda = 1, num.islands = 4, pop.size = 250,
  niter = 1000, parallel = TRUE)
}
\description{
Learn a linear dynamical system using Genetic Algorithm
}
