% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{LDS_reconstruction}
\alias{LDS_reconstruction}
\title{Learn LDS model with multiple initial conditions. This function can be called directly or wrapped in \link{LDS_ensemble}}
\usage{
LDS_reconstruction(
  Qa,
  u,
  v,
  start.year,
  method = "EM",
  trans = "log",
  init = NULL,
  num.restarts = 100,
  return.init = TRUE,
  lambda = 1,
  ub,
  lb,
  num.islands = 4,
  pop.per.island = 250,
  niter = 1000,
  tol = 1e-05,
  return.raw = FALSE,
  parallel = TRUE,
  all.cores = FALSE
)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{start.year}{Starting year of the paleo period. \code{start.year + ncol(u) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{method}{Either EM or GA (note: GA is experimental)}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{init}{A list of initial conditions, each element is a vector of length 4, the initial values for A, B, C and D. The initial values for Q and R are always 1, and mu_1 is 0 and V_1 is 1.}

\item{num.restarts}{if init is not given then num.restarts must be provided. In this case the function
will randomize the initial value by sampling uniformly within the range for each parameters
(A in [0, 1], B in [-1, 1], C in [0, 1] and D in [-1, 1]).}

\item{return.init}{Indicate whether the initial condition that results in the highest
log-likelihood is returned. Default is TRUE.}

\item{lambda}{Weight for penalty term (if method is not EM; experimental)}

\item{ub}{Upper bounds, a vector whose length is the number of parameters}

\item{lb}{Lower bounds}

\item{num.islands}{Number of islands (if method is GA; experimental)}

\item{pop.per.island}{Initial population per island (if method is GA; experimental)}

\item{niter}{Maximum number of iterations, default 1000}

\item{tol}{Tolerance for likelihood convergence, default 1e-5. Note that the log-likelihood is normalized by dividing by the number of observations.}

\item{return.raw}{If TRUE, state and streamflow estimates without measurement updates will be returned.}

\item{parallel}{If TRUE, the computation is done in parallel using all available cores}

\item{all.cores}{If TRUE, all cores are used, otherwise one less core will be used.
(using the doParallel backend). If FALSE, the computation is done serially.}
}
\value{
A list of the following elements
\itemize{
\item rec: reconstruction results, a data.table with the following columns
\itemize{
\item year: calculated from Qa and the length of u
\item X: the estimated hidden state
\item Xl, Xu: lower and upper range for the 95\\% confidence interval of X
\item Q: the reconstructed streamflow
\item Ql, Qu: lower and upper range for the 95\\% confidence interval of Q
}
\item theta: model parameters
\item lik: maximum likelihood
\item init: the initial condition that resulted in the maximum likelihood (if )
}
}
\description{
The initial conditions can either be randomized (specifiled by num.restarts) or provided beforehand.
}
