% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{ldsc_h2}
\alias{ldsc_h2}
\title{Estimate SNP heritability using LDscore regression}
\usage{
ldsc_h2(sumstat, weights = NULL, M = NULL)
}
\arguments{
\item{sumstat}{A data.frame or tbl with columns \code{SNP}, \code{Z} and \code{N}}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{M}{Optional, the number of SNPs in the reference panel}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{h2} and \code{h2_se}
}
\description{
An R implementation of the LD score regression method to estimate SNP heritability,
which should produce identical results to ldsc.py --h2 command.
The package uses by default the eur_w_ld scores from the LDSC tutorial.

The function does not apply any quality control checks, and assumes everything is in order.
}
\examples{
\dontrun{
ldsc_h2(my_gwas)
}
}
