% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_rg.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Compute the genetic correlation between two traits}
\usage{
ldsc_rg(sumstats1, sumstats2, weights = NULL, M = NULL, n_blocks = 200)
}
\arguments{
\item{sumstats1}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with atleast the columns SNP, A1, A2, Z, N.
To perform quality checks, use \code{\link[=munge]{munge()}} before running \code{ldsc_rg()}}

\item{sumstats2}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with atleast the columns SNP, A1, A2, Z, N.
To estimate rG with between sumstats1 and several traits, wrap the data.frames in a list.
If the names are lost, the column \code{trait2} will correspond to the index of the list.
Use a named list the retain the names in the \code{trait2} column
\code{list(dataframe1, dataframe2, ...)}}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{M}{Optional, the number of SNPs in the reference panel}

\item{n_blocks}{Number of blocks to use for the jackknife estimator}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Compute the genetic correlation between two traits
}
\examples{
path <- system.file("extdata", "eur_w_ld.parquet", package = "ldsR")
snps <- arrow::read_parquet(path, col_select = c("SNP"))
# to make example faster
snps <- dplyr::slice_head(snps, n = 100000)
snps$A1 <- "A"
snps$A2 <-
snps$N <- 130000
snps$Z <- rnorm(nrow(snps))
snps2 <- snps
snps2$N <- 75000
snps2$Z <- rnorm(nrow(snps))
ldsc_rg(snps, snps2)
# to run estimate the genetic correlations for many traits, wrap s2 in a list
# ldsc_rg(snps, list(snps2, snps2))
# use a named list to create the `trait2` column in the output
# ldsc_rg(snps, list("trait2" = s2, "trait3" = s3, "trait4" = s4))
}
