% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{ldsc_to_parquet}
\alias{ldsc_to_parquet}
\title{Transform a directory of LDscores to parquet format}
\usage{
ldsc_to_parquet(dir, thin = FALSE)
}
\arguments{
\item{dir}{directory containing the python LDSC ldscore format}

\item{thin}{If the --thin flag has been used, provide a character vector of RSIDs for the full dataset used to calculate LDscores}
}
\value{
a \code{\link[=list]{list()}}
}
\description{
Transform a directory of LDscores to parquet format
}
\examples{
\dontrun{
ldsc_to_parquet("/directory/ldsc")
}
}
