% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{ldsc_to_parquet2}
\alias{ldsc_to_parquet2}
\title{Transform LDSC formatted annotation ldscores to ldsR format}
\usage{
ldsc_to_parquet2(dir, outdir)
}
\arguments{
\item{dir}{directory with ldscores in LDSC format}

\item{outdir}{directory to save the parquet files}
}
\description{
Transform LDSC formatted annotation ldscores to ldsR format
}
\examples{
\dontrun{
ldsc_to_parquet2("path/to/ldsc", "path/to/outdir")
}
}
