% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{learnLDS}
\alias{learnLDS}
\title{Learn LDS model}
\usage{
learnLDS(y, u, v, init, niter, tol)
}
\arguments{
\item{y}{Observation matrix (may need to be normalized and centered before hand) (q rows, T columns)}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{mu}{The mean of y (if y was centralized before), to be added back to the prediction}
}
\value{
A list of fitted elements (X, Y, V, Cov, and lik)
\itemize{
\item X: a matrix of fitted states
\item Y: a matrix of fitted observation
\item V: a matrix of covariance of X
\item Cov: covariance of X_t and X_t-1
lik : log-likelihood
}
}
\description{
Estimate the hidden state and model parameters given observations and exogeneous inputs
}
\section{Note}{
 This code only works on one dimensional state and output at the moment. Therefore, transposing is skipped, and matrix inversion is treated as /, and log(det(Sigma)) is treated as log(Sigma).
}

