% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{learnLDS0}
\alias{learnLDS0}
\title{Learn LDS model}
\usage{
learnLDS0(y, u, v, init, niter, tol)
}
\description{
Estimate the hidden state and model parameters given observations and exogeneous inputs using the EM algorithm. This is the key backend routine of this package.
}
\details{
This is different from learnLDS in that here we assume perfect prediction.
This function is experimental, do not use.
}
