% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{learnLDS_restart}
\alias{learnLDS_restart}
\title{Learn LDS model with multiple initial conditions}
\usage{
learnLDS_restart(Qa, u, v, init = NULL, num.restart, niter = 1000,
  tol = 1e-05, return.init = TRUE, parallel = FALSE)
}
\arguments{
\item{Qa}{Observations: a data frame of annual streamflow with at least two columns: year and Qa.}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{init}{A list of initial conditions, each element is a vector of length 4, the initial values for A, B, C and D. The initial values for Q and R are always 1, and mu_1 is 0 and V_1 is 1.}

\item{num.restart}{if init is not given then num.restart must be provided. In this case the function will randomize the initial value by sampling uniformly within the range for each parameters (A in [0, 1], B in [-1, 1], C in [0, 1] and D in [-1, 1]).}

\item{niter}{Maximum number of iterations, default 1000}

\item{tol}{Tolerance for likelihood convergence, default 1e-5. Note that the log-likelihood is normalized by dividing by the number of observations.}

\item{return.init}{Indicate whether the initial condition that results in the highest log-likelihood is returned. Default is TRUE.}

\item{parallel}{If TRUE, the computation is done in parallel using all available cores (using the doParallel backend). If FALSE, the computation is done serially.}
}
\description{
The initial conditions can either be randomized (specifiled by num.restart) or provided beforehand.
}
