% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{make_init}
\alias{make_init}
\title{Make a list of initial parameter values for the search}
\usage{
make_init(init, d, num.restarts)
}
\arguments{
\item{init}{The init provided by the caller, can be NULL}

\item{d}{Dimension of input}

\item{num.restarts}{Number of randomized initial conditions}
}
\value{
A list of iniial conditions
}
\description{
If init is a vector, make it a list
If init is NULL, randomize it
}
