% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mstep}
\alias{Mstep}
\title{Maximizing expected likelihood using analytical solution}
\usage{
Mstep(y, u, v, fit)
}
\arguments{
\item{y}{Observation matrix (may need to be normalized and centered before hand) (q rows, T columns)}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{fit}{result of \link{Kalman_smoother}}
}
\value{
An object of class \code{theta}: a list of
}
\description{
Maximizing expected likelihood using analytical solution
}
