% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_GA.R
\name{negLogLik}
\alias{negLogLik}
\title{Calculates the negative log-likelihood}
\usage{
negLogLik(theta, u, v, y)
}
\arguments{
\item{theta}{A vector of model parameters, to be converted to a list}

\item{u}{Input matrix}

\item{v}{Input matrix}

\item{y}{Observations}
}
\value{
The negative log-likelihood
}
\description{
Calculates the negative log-likelihood
}
