% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPannual}
\alias{NPannual}
\title{Annual streamflow at Nakhon Phanom}
\format{
A data.table with two columns
\describe{
\item{year}{from 1960 to 2005}
\item{Qa}{mean annual streamflow, cubic meter per second}
}
}
\source{
Mekong River Commission
}
\usage{
NPannual
}
\description{
A dataset contaning annual streamflow of the Mekong River measured at station Nakhon Phanom.
}
\keyword{datasets}
