% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NPpc}
\alias{NPpc}
\title{Selected principal components}
\format{
A data.table
}
\usage{
NPpc
}
\description{
A dataset contaning the principal components of MADA surrounding NP
}
\keyword{datasets}
