% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR.R
\name{one_pcr_cv}
\alias{one_pcr_cv}
\title{One cross-validation run}
\usage{
one_pcr_cv(Xmat, y, z)
}
\arguments{
\item{Xmat}{The matrix of training data in the instrumental period.}

\item{y}{A vector of observations.}

\item{z}{A vector of left-out points.}
}
\value{
A vector of prediction.
}
\description{
Make one prediction for one cross-validation run. This is a subroutine to be called by other cross-validation functions.
}
