% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P1pc}
\alias{P1pc}
\title{Selected principal components}
\format{A data.table}
\usage{
P1pc
}
\description{
A dataset contaning the principal components of MADA surrounding P1).
}
\keyword{datasets}
