% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{parse_gwas}
\alias{parse_gwas}
\title{Parse a GWAS summary statistics}
\usage{
parse_gwas(df)
}
\arguments{
\item{df}{a in memory \code{\link[base:data.frame]{base::data.frame()}} or a filepath to ldsc-munged
summary statistics file or a filepath to a tidyGWAS folder.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{A1}, \code{A2}, \code{Z}, \code{N}
}
\description{
Parse a GWAS summary statistics
}
\examples{
\dontrun{
parse_gwas(tbl)
}
}
