% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{parse_parquet_dir}
\alias{parse_parquet_dir}
\title{Read LDscore format from ldsRs parquet format}
\usage{
parse_parquet_dir(dir, read_ref = FALSE)
}
\arguments{
\item{dir}{directory containing the files 'annot.parquet', 'annot_ref.parquet' and 'ld.parquet'}

\item{read_ref}{logical, whether to read the reference file 'annot_ref.parquet'}
}
\value{
a \code{\link[=list]{list()}}
}
\description{
Read LDscore format from ldsRs parquet format
}
\examples{
\dontrun{
files <- parse_parquet_dir("ldscores/atac")
}

}
