% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_h2.R
\name{partition_h2}
\alias{partition_h2}
\title{Flexible partitioning of heritability}
\usage{
partition_h2(
  sumstat,
  ldscore_dirs,
  subset_annots = NULL,
  overlapping_annotations = FALSE,
  weights = NULL,
  n_blocks = 200
)
}
\arguments{
\item{sumstat}{A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{Z} and \code{N}}

\item{ldscore_dirs}{a list of directories containing ldscore files}

\item{subset_annots}{a character vector of annotations to subset the ldscore files. Use \code{\link[=get_annot_names]{get_annot_names()}}}

\item{overlapping_annotations}{are the annotations overlapping? In such a case, the estimate of the enrichment estimate needs to be adjusted.}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{n_blocks}{Number of blocks to use for the jackknife estimator}
}
\value{
a tibble with results
}
\description{
An R implementation of the LD score regression method to estimate SNP heritability, focusing on partioning heritability across multiple annotations.
}
\examples{
\dontrun{
partition_h2(sumstat, ldscore_dirs, subset_annots = c("L2", "L3"))
}
}
