% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{PCR_ensemble}
\alias{PCR_ensemble}
\title{Same as \code{PCR} but using ensemble model}
\usage{
PCR_ensemble(Qa, pc.list, k, CV.reps = 100, Z = NULL)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc.list}{A list, each element is a set of principal component as in \code{PCR_reconstruction}'s \code{pc}}

\item{k}{Numer of data points to be left out in each cross validation-run.}

\item{CV.reps}{Number of cross-validation runs.}

\item{Z}{A list of CV.reps elements, each is a vector of length k. See Details.}
}
\description{
Same as \code{PCR} but using ensemble model
}
