% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{PCR_reconstruction}
\alias{PCR_reconstruction}
\title{Reconstruction with Principal Component Regression}
\usage{
PCR_reconstruction(Qa, pc, start.year, trans = "log", stepwise = TRUE)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc}{A data frame, one colulmn for each principal component}

\item{start.year}{Starting year of \code{pc}, i.e, the first year of the paleo period. \code{start.year + nrow(pc) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{stepwise}{If \code{TRUE}, backward stepwise selection will be performed. Otherwise, all PCs will be used.}
}
\value{
A list of reconstruction results
\itemize{
\item rec: reconstruction
\item selected: a vector of selected principal components
}
}
\description{
Reconstruct streamflow using the given principal compone11nts and backward selection.
Since this is the benchmark for the LDS model, and for simplicity, learning and cross-validation are performed in one go, instead of with two separate functions as in the case of LDS.
}
