% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_GA.R
\name{penalized_likelihood}
\alias{penalized_likelihood}
\title{Penalized likelihood objective function}
\usage{
penalized_likelihood(y, u, v, theta.vec, lambda)
}
\arguments{
\item{y}{Observation matrix (may need to be normalized and centered before hand) (q rows, T columns)}

\item{u}{Input matrix for the state equation (m_u rows, T columns)}

\item{v}{Input matrix for the output equation (m_v rows, T columns)}

\item{theta.vec}{a vector of parameter elements (i.e, the vectorized version of theta
in \code{Kalman_smoother})}

\item{lambda}{weight of the penalty}
}
\value{
The penalized likelihood (a real number)
}
\description{
\link{Kalman_smoother} returns X and likelihood.
The penalized likelihood is the likelihood minus the sum-of-squares of the
measurement update. This is used as the fitness function in genetic algorihm.
}
