% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ping.annual}
\alias{Ping.annual}
\title{P1 annual streamflow}
\format{A data frame with 85 rows and 2 variables:
\describe{
  \item{year}{from 1921 to 1985}
  \item{Qa}{annual streamflow in million cubic meters per month}
}}
\source{
\url{https://www.hydro-1.net/08HYDRO/HD-04/4-06.html}
}
\usage{
Ping.annual
}
\description{
A dataset contaning annual streamflow of the Ping River measured at station P1 (Chiang Mai,
Thailand).
}
\keyword{datasets}
