% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ping.PC}
\alias{Ping.PC}
\title{Selected principal components}
\format{A matrix of 7 rows and 406 columns}
\usage{
Ping.PC
}
\description{
A dataset contaning the principal components of MADA surrounding P1).
}
\keyword{datasets}
