% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_functions.R
\name{plot_reconstruction}
\alias{plot_reconstruction}
\title{Plot reconstruction results in the instrumental period or full period}
\usage{
plot_reconstruction(s, target, period)
}
\arguments{
\item{s}{Reconstruction results for a site}

\item{target}{Reconstruction target}
}
\description{
Plot reconstruction results in the instrumental period or full period
}
