% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastics.R
\name{plot_replicates}
\alias{plot_replicates}
\title{Plot stochastic replicates}
\usage{
plot_replicates(reps)
}
\arguments{
\item{reps}{Results of \link{LDS_rep}.}
}
\value{
A plot with two panels: stochastic replicates of streamflow and catchment state
}
\description{
Plot stochastic replicates
}
