% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RE}
\alias{RE}
\title{Reduction of Error}
\usage{
RE(yhat, y, yc_bar)
}
\arguments{
\item{yhat}{Model outputs in the validation set}

\item{y}{Observations in the validation set}

\item{yc_bar}{Mean observations in the calibration set}
}
\value{
RE
}
\description{
Reduction of Error
}
