% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theta}
\alias{theta}
\title{LDS parameters}
\format{
A list with elements A, B, C, D, Q, R, mu1, V1
}
\usage{
theta
}
\description{
Theta values for Nakhon Phanom, precomputed to use in examples.
}
\keyword{datasets}
