% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{to_celltype_dataset}
\alias{to_celltype_dataset}
\title{Convert the output of many LDscores to a single dataframe suitable for ldsR}
\usage{
to_celltype_dataset(parent_dir, outdir, ref_snps = NULL)
}
\arguments{
\item{parent_dir}{a directory with subdirectories containing LDscore data}

\item{outdir}{directory to save the parquet files}

\item{ref_snps}{a character vector of RSIDs corresponding to SNPs used in the
full dataset used to generate LDscores. Provide if --thin-annot has been used
in the LDSC command to generate ldscores}
}
\description{
Convert the output of many LDscores to a single dataframe suitable for ldsR
}
\examples{
\dontrun{
to_celltype_dataset("files/ldsc", "files/ldsc_parquet")
}
}
