% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_GA.R
\name{vec_to_list}
\alias{vec_to_list}
\title{Converts theta from a vector (as used in GA) to list (as used in Kalman smoothing)}
\usage{
vec_to_list(theta.vec, d)
}
\arguments{
\item{theta.vec}{a vector of parameter elements}

\item{d}{dimention of inputs}
}
\value{
A theta object, see \link{make_init}
}
\description{
Converts theta from a vector (as used in GA) to list (as used in Kalman smoothing)
}
