% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{water_to_calendar_year}
\alias{water_to_calendar_year}
\title{Water to calendar year}
\usage{
water_to_calendar_year(df, dt, keep.all = FALSE)
}
\arguments{
\item{df}{A data frame following water year with at least two columns
\describe{
\item{year}{The water year}
\item{months}{The months in the water year, with \code{month = 1} indicating the first month of the water year}
}}

\item{dt}{The time shift required to change from water year to calendar year. For example, if the water year 2000 starts from April 2000 then \code{delta.t = 3} (caldendar year is 3 months a head of water year). On the other hand, if the water year 2000 starts from October 1999 then \code{delta.t = -3} (calendear year is 3 months behind water year).}

\item{keep.all}{If \code{FALSE} (the default), only complete years are retained. Otherwise, the whole data frame is retained.}
}
\value{
A new data frame where the year and month is now shifted to calendar year, with \code{month = 1} indicating January.
}
\description{
Converting a sub-annual data frame from water year to calendar year. Users must pay attention to the notion of the water year, whether it uses the year in which it starts (e.g., water year 2000 is from April 2000 to March 2001) or the year in which it ends (e.g., water year 2000 is from September 1999 to October 2000).
}
\details{
For example, in Thailand a water year is from April to March, i.e, month 1 of 1921 denotes April 1921, and month 12 of 1921 denotes March 1922. After shifting, the period of April to December 1921 are discarded, and the data frame starts from January 1922.
}
