% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype-analysis.R
\name{celltype_analysis}
\alias{celltype_analysis}
\title{Perform cell-type analysis using partitioned heritability}
\usage{
celltype_analysis(sumstat, covariate_dir, ldscore_dir, weights = NULL)
}
\arguments{
\item{sumstat}{A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{Z} and \code{N}}

\item{covariate_dir}{a directory containing the files \code{ld.parquet} and \code{annot.parquet}}

\item{ldscore_dir}{filepath to a directory with the \code{annot.parquet} and \code{ldscores.parquet} files}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
A common usage of partitioned heritability is to estimate enrichment and significance for
an annotation typical for a specific cell-type.
For this analysis, it is useful to adjust for a baseline set of annotations, and then estimate the association for a large set of annotations ("cell-types")
}
\examples{
\dontrun{
sumstat1 <- arrow::read_parquet(system.file("extdata", "sumstats.parquet", package = "ldsR"))
sumstat1 <- dplyr::select(sumstat1, SNP, Z = Z.x, N=N.x)


celltype_analysis(
 sumstat1,
 covariate_dir = system.file("extdata", "baseline1.1_test", package = "ldsR"),
 ldscore_dir = system.file("extdata", "superclusters", package = "ldsR")
)
}

}
