% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-analysis.R
\name{celltype_analysis}
\alias{celltype_analysis}
\title{Run partitioned heritability across many annotations}
\usage{
celltype_analysis(sumstat, covariate_dir, celltype_dir, weights = NULL)
}
\arguments{
\item{sumstat}{A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{Z} and \code{N}}

\item{covariate_dir}{a directory containing the files \code{ld.parquet} and \code{annot.parquet}}

\item{celltype_dir}{a directory containing the files \code{ld.parquet} and \code{annot.parquet}}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Run partitioned heritability across many annotations
}
\examples{
\dontrun{
celltype_analysis(ss_tbl, "/path/to_baseline", "path/to/celltype_ldscores")
}
}
