% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-analysis.R
\name{celltype_analysis}
\alias{celltype_analysis}
\title{Run partitioned heritability across many annotations}
\usage{
celltype_analysis(sumstat, covariate_dir, celltype_dir, weights = NULL)
}
\arguments{
\item{sumstat}{A data.frame or tbl with columns `SNP`, `Z` and `N`}

\item{covariate_dir}{a directory containing the files `ld.parquet` and `annot.parquet`}

\item{celltype_dir}{a directory containing the files `ld.parquet` and `annot.parquet`}

\item{weights}{A numeric vector of regression weights. Comes precomputed within the ldsR package for european LDscores}
}
\value{
a [dplyr::tibble()]
}
\description{
Run partitioned heritability across many annotations
}
\examples{
\dontrun{
celltype_analysis(ss_tbl, "/path/to_baseline", "path/to/celltype_ldscores")
}
}
