% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\name{combine_ld_data}
\alias{combine_ld_data}
\title{Combine two ldsc_to_parquet outputs}
\usage{
combine_ld_data(list1, list2, ref = TRUE, outdir = NULL)
}
\arguments{
\item{list1, }{list2 output of ldsc_to_parquet}

\item{outdir}{if NULL, return the combined data, otherwise save to outdir}
}
\value{
a list or NULL
}
\description{
Combine two ldsc_to_parquet outputs
}
\examples{
\dontrun{
combine_ld_data("files/ldsc_parquet/CD4_T_cells", "files/ldsc_parquet/CD8_T_cells")
}
}
