% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_boxcox}
\alias{inv_boxcox}
\title{Inverse Box-Cox transform}
\usage{
inv_boxcox(x, lambda)
}
\arguments{
\item{x}{A numeric vector to be transformed}

\item{lambda}{Lambda parameter}
}
\value{
The inverse Box-Cox
}
\description{
Inverse Box-Cox transform
}
\examples{
inv_boxcox(x = rnorm(10), lambda = 1)
inv_boxcox(x = rnorm(10), lambda = 0) # exp(x)
}
