% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_GA.R
\name{LDS_BFGS}
\alias{LDS_BFGS}
\title{Learn LDS with L-BFGS-B}
\usage{
LDS_BFGS(y, u, v, ub, lb, num.restarts = 100, parallel = TRUE)
}
\arguments{
\item{y}{Transformed and standardized streamflow}

\item{u}{Input matrix for a single-model reconstruction, or a list of input matrices for an ensemble reconstruction.}

\item{v}{Same as u.}

\item{ub}{Upper bounds, a vector whose length is the number of parameters}

\item{lb}{Lower bounds}

\item{num.restarts}{The number of initial conditions to start the EM search; ignored if \code{init} is provided.}

\item{parallel}{Logical, whether parallel computation is used}
}
\value{
A list of reconstruction results; see \link{LDS_reconstruction}
}
\description{
\strong{Warning} This is an experimental feature. Use with care.
}
