% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{ldsc_h2}
\alias{ldsc_h2}
\title{Estimate SNP heritability using LDscore regression}
\usage{
ldsc_h2(sumstat, weights = NULL, M = NULL, n_blocks = 200)
}
\arguments{
\item{sumstat}{A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{Z} and \code{N}}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{M}{Optional, the number of SNPs in the reference panel}

\item{n_blocks}{Number of blocks to use for the jackknife estimator}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{h2} and \code{h2_se}
}
\description{
An R implementation of the LD score regression method to estimate SNP heritability,
mimicking \code{ldsc --h2} from the \href{https://github.com/bulik/ldsc/wiki/Heritability-and-Genetic-Correlation}{ldsc} package.

LDscores for the European subset of 1000g (2015 release) for 1,290,028 HapMap3 SNPs (with 1,173,569 SNPs with freq > 5\%)
are bundled within the ldsR package and are used by default. This corresponds to the \code{eur_w_ld_chr} folder
previously shared at the LDSC \href{https://github.com/bulik/ldsc}{github}.

You can inspect the LDscores used by default:
\code{arrow::read_parquet(system.file("extdata", "eur_w_ld.parquet", package = "ldsR"))}

ldsc_h2 does not perform any quality control on the input summary statistics, except to merge with the
1,290,028 HapMap3 SNPs in the reference panel. See the \code{\link[=munge]{munge()}} function to mimic the \code{munge_sumstats.py} function.
}
\examples{
p <- system.file("extdata", "eur_w_ld.parquet", package = "ldsR")
snps <- arrow::read_parquet(p, col_select = c("SNP"))
snps$N <- 130000
snps$Z <- rnorm(nrow(snps))
ldsc_h2(snps)


}
