% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_rg.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Compute the genetic correlation between two traits}
\usage{
ldsc_rg(sumstats1, sumstats2, weights = NULL, M = NULL, n_blocks = 200)
}
\arguments{
\item{sumstats1}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with atleast the columns SNP, A1, A2, Z, N}

\item{sumstats2}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with atleast the columns SNP, A1, A2, Z, N.
A list of summary statistics can be passed to estimate rG between sumstats1 and all traits in sumstats2}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{M}{Optional, the number of SNPs in the reference panel}

\item{n_blocks}{Number of blocks to use for the jackknife estimator}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Compute the genetic correlation between two traits
}
\examples{
\dontrun{
ldsc_rg(s1, s2)
}
}
