% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsr.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Compute the genetic correlation between two traits}
\usage{
ldsc_rg(sumstats1, sumstats2, weights = NULL, M = NULL, n_blocks = 200)
}
\arguments{
\item{sumstats1}{a [dplyr::tibble()] with columns h2 and h2_se}

\item{sumstats2}{a [dplyr::tibble()] with columns h2 and h2_se}

\item{weights}{Optional, a data.frame or tbl with columns SNP and L2}

\item{M}{number of snps used to calculate LD in weights}

\item{n_blocks}{blocks for jackknife}
}
\value{
a [dplyr::tibble()]
}
\description{
Compute the genetic correlation between two traits
}
