% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsr-package.R
\docType{package}
\name{ldsr}
\alias{ldsr}
\alias{ldsr-package}
\title{ldsr: Linear Dynamical System Reconstruction}
\description{
Streamflow (and climate) reconstruction using Linear Dynamical Systems. 
    The advantage of this method is the additional state trajectory which can reveal more information 
    about the catchment or climate system. For details of the method please refer to Nguyen and Galelli 
    (2018) <doi:10.1002/2017WR022114>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ntthung/ldsr}
  \item Report bugs at \url{https://github.com/ntthung/ldsr/issues}
}

}
\author{
\strong{Maintainer}: Hung Nguyen \email{ntthung@gmail.com}

Other contributors:
\itemize{
  \item Stefano Galelli \email{stefano_galelli@sutd.edu.sg} [contributor]
}

}
\keyword{internal}
