% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsr.R
\docType{package}
\name{ldsr}
\alias{ldsr}
\alias{ldsr-package}
\title{ldsr: Paleo reconstruction with linear dynamical systems}
\description{
Reconstruct paleoclimate variables (streamflow, precipitation, etc...) with linear dynamical systems
}
\author{
\strong{Maintainer}: Hung Nguyen \email{ntthung@gmail.com}

Authors:
\itemize{
  \item Stefano Galelli \email{stefano_galelli@sutd.edu.sg}
}

}
\keyword{internal}
