% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDS_reconstruction.R
\name{make_init}
\alias{make_init}
\title{Make initial values for parameters.}
\usage{
make_init(p, q, num.restarts)
}
\arguments{
\item{p}{Dimension of u}

\item{q}{Dimension of v}

\item{num.restarts}{Number of randomized initial values}
}
\value{
A list of initial conditions, each element is an object of class \code{theta}.
}
\description{
If init is a vector, make it a list of one element.
If init is NULL, randomize it. In this case, the function will randomize the initial value by sampling uniformly within the range for each parameters
(A in [0, 1], B in [-1, 1], C in [0, 1] and D in [-1, 1]).
}
\examples{
make_init(5, 5, 1)
make_init(5, 5, 2)
}
