% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{munge}
\alias{munge}
\title{Munge GWAS summary statistics}
\usage{
munge(dset, info_filter = 0.9, eaf_filter = 0.01)
}
\arguments{
\item{dset}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{A1} \code{A2} \code{Z} \code{N} and
optional columns \code{EAF} and \code{INFO}}

\item{info_filter}{INFO score filter threshold at which to remove rows}

\item{eaf_filter}{effective allele filter at which to remove rows.
eaf_filter=0.01 would filter to eaf > 0.01 & eaf < 0.99}
}
\value{
a data.frame
}
\description{
\code{munge()} applies 5 filters (if possible):
\enumerate{
\item Deduplication of RSID
\item Filter variants on INFO in INFO column is present
\item Filter variant on effect allele frequency
\item Removes strand ambigious variants
\item Removes variants with \code{N < round(stats::quantile(N, 0.9) / 1.5)}
}
}
\examples{
\dontrun{
munge(tbl)
}
}
