% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{munge}
\alias{munge}
\title{Munge GWAS summary statistics}
\usage{
munge(dset, info_filter = 0.9, maf_filter = 0.01)
}
\arguments{
\item{dset}{a \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{A1} \code{A2} \code{Z} \code{N} and possibly \code{EAF} and \code{INFO}}

\item{info_filter}{INFO score filter threshold at which to remove rows}

\item{maf_filter}{Minor allele frequenc filter at which to remove rows}
}
\value{
a data.frame
}
\description{
Munge GWAS summary statistics
}
\examples{
\dontrun{
parse_gwas(tbl)
}
}
