% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nRMSE}
\alias{nRMSE}
\title{Normalized root-mean-square error}
\usage{
nRMSE(yhat, y, normConst)
}
\arguments{
\item{yhat}{Model outputs}

\item{y}{Observations}

\item{normConst}{The normalization constant}
}
\value{
normalized RMSE
}
\description{
RMSE is normalized by the normalization constant
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
nRMSE(x, y, sd(y))
}
