% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{parse_gwas}
\alias{parse_gwas}
\title{Parse a GWAS summary statistics}
\usage{
parse_gwas(df, format = c("dataframe", "ldsc", "tidyGWAS"))
}
\arguments{
\item{df}{a in memory \code{\link[base:data.frame]{base::data.frame()}} or a filepath to ldsc-munged
summary statistics file or a filepath to a tidyGWAS folder.}

\item{format}{format of summary statistics. Either a in-memory dataframe, and LDSC munged file, or tidyGWAS filepath}
}
\value{
a data.frame
}
\description{
Parse a GWAS summary statistics
}
\examples{
\dontrun{
parse_gwas(tbl)
}
}
