% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned_h2.R
\name{partitioned_h2}
\alias{partitioned_h2}
\title{Estimate partitioned SNP heritability}
\usage{
partitioned_h2(
  sumstat,
  ldscore_dir,
  weights = NULL,
  n_blocks = 200,
  overlapping_annotations = NULL
)
}
\arguments{
\item{sumstat}{A \code{\link[dplyr:reexports]{dplyr::tibble()}} with columns \code{SNP}, \code{Z} and \code{N}}

\item{ldscore_dir}{filepath to a directory with the \code{annot.parquet} and \code{ldscores.parquet} files}

\item{weights}{Optional, a data.frame or tbl with columns \code{SNP}, \code{L2}}

\item{n_blocks}{Number of blocks to use for the jackknife estimator}

\item{overlapping_annotations}{are the annotations overlapping? In such a case, the estimate of the enrichment estimate needs to be adjusted.}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
An R implementation of the LD score regression method to estimate SNP heritability,
mimicking \code{ldsc --h2} from the \href{https://github.com/bulik/ldsc/wiki/Heritability-and-Genetic-Correlation}{ldsc} package.

For partitioned heritablity, the noMHC weights are used instead of the eur_w_ld_chr weights.

You can inspect the LDscores used by default, in the column \code{L2_celltype}
\code{arrow::read_parquet(system.file("extdata", "eur_w_ld.parquet", package = "ldsR"))}

partitioned_heritability does not perform any quality control on the input summary statistics, except to merge with the
SNPs in the reference panel. See the \code{\link[=munge]{munge()}} function to mimic the \code{munge_sumstats.py} function.
}
\examples{
\dontrun{
sumstat1 <- arrow::read_parquet(system.file("extdata", "sumstats.parquet", package = "ldsR")) |>
dplyr::select(SNP, Z = Z.x, N=N.x)
ldscore_dir <-  system.file("extdata", "baseline1.1_test", package = "ldsR")
fs::dir_tree(ldscore_dir)
partitioned_h2(sumstat1, ldscore_dir = ldscore_dir)
}



}
