% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{PCR_ensemble}
\alias{PCR_ensemble}
\title{Same as \code{PCR} but using ensemble model}
\usage{
PCR_ensemble(Qa, pc.list, start.year, trans = "log", stepwise = TRUE)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc.list}{A list, each element is a set of principal component as in \code{PCR_reconstruction}'s \code{pc}}

\item{start.year}{Starting year of \code{pc}, i.e, the first year of the paleo period. \code{start.year + nrow(pc) - 1} will determine the last year of the study horizon, which must be greater than or equal to the last year in \code{Qa}.}

\item{trans}{Transformatin of flow before model fitting, can be either "log", "boxcox" or "none".}

\item{stepwise}{If \code{TRUE}, backward stepwise selection will be performed. Otherwise, all PCs will be used.}
}
\description{
Same as \code{PCR} but using ensemble model
}
