% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{PCR_reconstruction}
\alias{PCR_reconstruction}
\title{Reconstruction with Principal Component Regression}
\usage{
PCR_reconstruction(Qa, pc, k, CV.reps = 100, Z = NULL)
}
\arguments{
\item{Qa}{Observations: a data.table of annual streamflow with at least two columns: year and Qa.}

\item{pc}{A data frame of principal components with first column called years (the years in) the
study period, followed by one column for each principal components.}

\item{k}{Numer of data points to be left out in each cross validation-run.}

\item{CV.reps}{Number of cross-validation runs.}

\item{Z}{A list of CV.reps elements, each is a vector of length k. See Details.}
}
\value{
A list of reconstruction and cross-validation results
\itemize{
\item rec: reconstruction
\item metrics: performance metrics
\item metrics.dist: distribution of performance metrics across all cross-validation runs.
\item Ycv: the predicted streamflow in each cross validation run
\item Z: the cross-validation points
}
}
\description{
Reconstruct streamflow using the given principal components and backward selection.
Since this is the benchmark for the LDS model, and for simplicity, learning and cross-validation are performed in one go, instead of with two separate functions as in the case of LDS.
}
